refactor text formatting to helper class
this makes it possible to test chat heads
